backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.5
pillow==9.5.0
requests==2.28.2
jinja2==3.1.2
